requirements.txt
mysql.connector
uvicorn
fastapi~=0.116.1
